test: integrations: nrt: extract helpers
extract common code to be used also in the upcoming
integration tests for the cache/reserve functionality.

(cherry picked from commit 698586bb82316fdba811155a9499231a5e01fbbf)
